* server update. * package dependencies update. * remove appveyor.
