+++ package.json
@@ -40,7 +40,6 @@
     "browser-sync": "^2.13.0",
     "codelyzer": "0.0.25",
     "colorguard": "^1.2.0",
-    "compression": "^1.6.2",
     "connect": "^3.4.1",
     "connect-history-api-fallback": "^1.2.0",
     "connect-livereload": "^0.5.4",
@@ -120,8 +119,10 @@
     "@angular/platform-browser": "2.0.0-rc.4",
     "@angular/platform-browser-dynamic": "2.0.0-rc.4",
     "@angular/router": "3.0.0-beta.2",
+    "compression": "^1.6.2",
     "core-js": "^2.4.0",
     "es6-module-loader": "^0.17.8",
+    "express": "^4.14.0",
     "pm2": "^1.1.3",
     "rxjs": "5.0.0-beta.6",
     "systemjs": "0.19.27",
